Refactoring nav bar
